requirements.txt
fastapi
uvicorn
langchain
langchain-community
pydantic
python-dotenv
openai>=1.0.0
google-api-python-client

